Implement a select box (#73)
### Issue

There is an issue that A `TouchableOpacity` in `absolute` view can not be clickable on `android`. I posted this issue [here](https://stackoverflow.com/questions/68916203/react-native-touchableopacity-in-absolute-view-can-not-be-clickable-on-andro).

### Update
When I set `pointerEvents` property to `box-none` to the parent view of `TouchableOpacity`, then touch event can reach to the right place.

But It is quite tricky, and can not works with `zIndex`. So it's hard to show dropdown over other component on `android`. but I'm getting closer to the solution.

Co-authored-by: hyochan <dooboolab@gmail.com>