refactor: change button to [TouchableHighlight] with overlay
The [Button] is now wrapped with [TouchableHighlight] instead of [TouchableOpacity] and it needs to pass `onPress` props to react to press unlike [TouchableOpacity]. Thus, I removed `borderRadius` and `padding` from `style` in typescript to let user know that it will cause unexpected behavior. When `padding` is given to [TouchableHightlight], it will have `underlayColor` as extra background. In this case, you can rather use `margin`.