refactor: [Button] component and organize stories  (#231)
Change [Button] type from
[TouchableOpacity](https://reactnative.dev/docs/touchableopacity) to
[TouchableHighlight](https://reactnative.dev/docs/touchablehighlight).
This decision was made because removing valid color with `opacity` when
pressing does not feel like the Button although this is easier to use.
We rather choose to give darker color when `pressIn`. With this
decision, we should not use `padding` and `borderRadius` in `style`
props because there is an underlying overlay component in
`TouchableHighlight` so the [Button] will have `underlayColor` as a
`backgroundColor`. These style props are omitted in `typescript` for
safety.

- Organize [Button] stories by `title` for readability.
- Make `commitlint` have a `maxLength` of 500 for the body.
- Change the `render` function to an element that does not need to be a
function.
- Add test for text button and remove unreached code
- Fix [Button] test not to repeat the common props and theme type.
Previous code makes tests hard to fix since all test components were
copied and pasted.
- Add test for text button and remove unreached code.