refactor: stories (#236)
Refactor stories by managing all stories in the same container which is
wrapped with [DoobooProvider]. Also, moved `stories` to a separate
`stories` directory to use common components which are used only in
stories. We can expect common stories components to be used in `docs`
either.

Combine `dark` and the `light` mode in a single story. You can now use
[SwitchToggle] to switch themes in stories.

- chore: update icons in `ttf`