refactor: remove useButtonKnobs and manage const for stories (#242)
Remove `useButtonKnobs` and include global `const` for stories.

Previously, `useButtonKnobs` was used to be used in similar stories in
[Button]. However, I face a problem when some button stories do not
include knobs defined in `useButtonKnobs`. Knobs that are not used in
the story are also revealed because they are defined in
`useButtonKnobs`. Therefore, I've removed `useButtonKnobs` and sadly
copy and pasted the same codes in button stories.

Add `const.ts` file for managing const variables used in
stories.